feat: refactor product DTOs and mappers for improved data handling and consistency

+++ src/main/java/com/example/tileshop/dto/product/ProductDetailResponseDTO.java
@@ -27,7 +27,7 @@ public class ProductDetailResponseDTO {
 
     private double price;
 
-    private Double discountPercentage;
+    private int discountPercentage;
 
     private double salePrice;
